Create README.md
A Led blinking is a basic program. In this article, we will learn how to interface led with 8051 micro-controller. The 8051 is basic micro-controller and it’s the best for a beginner. In 8051 when we have turned the power on, all I/O pins have the high value. Means that initially all I/O pins work as input except PORT 0 because in 8051 all ports have the inbuilt pull-up resistor except PORT 0.